Relocate CLI git interactions
