+++ lib/python/qmk/commands.py
@@ -4,13 +4,12 @@ import os
 import sys
 import shutil
 from pathlib import Path
-from subprocess import DEVNULL
 
 from milc import cli
 import jsonschema
 
 import qmk.keymap
-from qmk.constants import QMK_FIRMWARE, KEYBOARD_OUTPUT_PREFIX
+from qmk.constants import KEYBOARD_OUTPUT_PREFIX
 from qmk.json_schema import json_load, validate
 
 
@@ -88,31 +87,6 @@ def create_make_command(keyboard, keymap, target=None, dry_run=False, parallel=1
         make_args.append(target)
 
     return create_make_target(':'.join(make_args), dry_run=dry_run, parallel=parallel, **env_vars)
-
-
-def get_git_version(current_time, repo_dir='.', check_dir='.'):
-    """Returns the current git version for a repo, or the current time.
-    """
-    git_describe_cmd = ['git', 'describe', '--abbrev=6', '--dirty', '--always', '--tags']
-
-    if repo_dir != '.':
-        repo_dir = Path('lib') / repo_dir
-
-    if check_dir != '.':
-        check_dir = repo_dir / check_dir
-
-    if Path(check_dir).exists():
-        git_describe = cli.run(git_describe_cmd, stdin=DEVNULL, cwd=repo_dir)
-
-        if git_describe.returncode == 0:
-            return git_describe.stdout.strip()
-
-        else:
-            cli.log.warn(f'"{" ".join(git_describe_cmd)}" returned error code {git_describe.returncode}')
-            print(git_describe.stderr)
-            return current_time
-
-    return current_time
 
 
 def get_make_parallel_args(parallel=1):
@@ -235,83 +209,6 @@ def parse_configurator_json(configurator_file):
             user_keymap['layout'] = aliases[orig_keyboard]['layouts'][user_keymap['layout']]
 
     return user_keymap
-
-
-def git_get_username():
-    """Retrieves user's username from Git config, if set.
-    """
-    git_username = cli.run(['git', 'config', '--get', 'user.name'])
-
-    if git_username.returncode == 0 and git_username.stdout:
-        return git_username.stdout.strip()
-
-
-def git_check_repo():
-    """Checks that the .git directory exists inside QMK_HOME.
-
-    This is a decent enough indicator that the qmk_firmware directory is a
-    proper Git repository, rather than a .zip download from GitHub.
-    """
-    dot_git_dir = QMK_FIRMWARE / '.git'
-
-    return dot_git_dir.is_dir()
-
-
-def git_get_branch():
-    """Returns the current branch for a repo, or None.
-    """
-    git_branch = cli.run(['git', 'branch', '--show-current'])
-    if not git_branch.returncode != 0 or not git_branch.stdout:
-        # Workaround for Git pre-2.22
-        git_branch = cli.run(['git', 'rev-parse', '--abbrev-ref', 'HEAD'])
-
-    if git_branch.returncode == 0:
-        return git_branch.stdout.strip()
-
-
-def git_get_tag():
-    """Returns the current tag for a repo, or None.
-    """
-    git_tag = cli.run(['git', 'describe', '--abbrev=0', '--tags'])
-    if git_tag.returncode == 0:
-        return git_tag.stdout.strip()
-
-
-def git_is_dirty():
-    """Returns 1 if repo is dirty, or 0 if clean
-    """
-    git_diff_staged_cmd = ['git', 'diff', '--quiet']
-    git_diff_unstaged_cmd = [*git_diff_staged_cmd, '--cached']
-
-    unstaged = cli.run(git_diff_staged_cmd)
-    staged = cli.run(git_diff_unstaged_cmd)
-
-    return unstaged.returncode != 0 or staged.returncode != 0
-
-
-def git_get_remotes():
-    """Returns the current remotes for a repo.
-    """
-    remotes = {}
-
-    git_remote_show_cmd = ['git', 'remote', 'show']
-    git_remote_get_cmd = ['git', 'remote', 'get-url']
-
-    git_remote_show = cli.run(git_remote_show_cmd)
-    if git_remote_show.returncode == 0:
-        for name in git_remote_show.stdout.splitlines():
-            git_remote_name = cli.run([*git_remote_get_cmd, name])
-            remotes[name.strip()] = {"url": git_remote_name.stdout.strip()}
-
-    return remotes
-
-
-def git_check_deviation(active_branch):
-    """Return True if branch has custom commits
-    """
-    cli.run(['git', 'fetch', 'upstream', active_branch])
-    deviations = cli.run(['git', '--no-pager', 'log', f'upstream/{active_branch}...{active_branch}'])
-    return bool(deviations.returncode)
 
 
 def in_virtualenv():
